+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/util/SourceUtil.java
@@ -14,8 +14,8 @@
 
 package com.liferay.source.formatter.checks.util;
 
+import com.liferay.petra.string.CharPool;
 import com.liferay.portal.kernel.io.unsync.UnsyncStringReader;
-import com.liferay.portal.kernel.util.CharPool;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -119,7 +119,7 @@ public class SourceUtil {
 		for (String line : lines) {
 			line = StringUtil.trim(line);
 
-			if (line.startsWith("//")) {
+			if (line.startsWith("//") || line.startsWith("*")) {
 				continue;
 			}