+++ portal-impl/src/com/liferay/portal/upgrade/v7_1_0/UpgradeModules.java
@@ -31,8 +31,10 @@ public class UpgradeModules
 	}
 
 	private static final String[] _BUNDLE_SYMBOLIC_NAMES = {
+		"com.liferay.document.library.content.service",
 		"com.liferay.document.library.file.rank.service",
 		"com.liferay.document.library.sync.service",
+		"com.liferay.message.boards.service",
 		"com.liferay.subscription.service", "com.liferay.trash.service"
 	};