+++ modules/sdk/gradle-plugins/CHANGELOG.markdown
@@ -2748,6 +2748,30 @@ to version 2.3.15.
 - [LPS-75323]: Update the [Liferay Source Formatter] dependency to version
 1.0.486.
 
+## 3.5.24 - 2017-10-31
+
+### Changed
+- [LPS-75488]: Update the [Liferay Gradle Plugins Source Formatter] dependency
+to version 2.3.16.
+- [LPS-75488]: Update the [Liferay Source Formatter] dependency to version
+1.0.487.
+
+## 3.5.25 - 2017-11-01
+
+### Changed
+- [LPS-75613]: Update the [Liferay Gradle Plugins Source Formatter] dependency
+to version 2.3.17.
+- [LPS-75613]: Update the [Liferay Source Formatter] dependency to version
+1.0.488.
+
+## 3.5.26 - 2017-11-01
+
+### Changed
+- [LPS-75589]: Update the [Liferay CSS Builder] dependency to version 2.0.1.
+- [LPS-75589]: Update the [Liferay Gradle Plugins CSS Builder] dependency to
+version 2.1.6.
+- [LPS-75624]: Set the default Node.js version to 8.9.0.
+
 [Gradle Bundle Plugin]: https://github.com/TomDmitriev/gradle-bundle-plugin
 [Liferay Alloy Taglib]: https://github.com/liferay/alloy-taglibs
 [Liferay Ant BND]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/ant-bnd
@@ -2995,4 +3019,8 @@ to version 2.3.15.
 [LPS-75323]: https://issues.liferay.com/browse/LPS-75323
 [LPS-75427]: https://issues.liferay.com/browse/LPS-75427
 [LPS-75430]: https://issues.liferay.com/browse/LPS-75430
+[LPS-75488]: https://issues.liferay.com/browse/LPS-75488
+[LPS-75589]: https://issues.liferay.com/browse/LPS-75589
+[LPS-75613]: https://issues.liferay.com/browse/LPS-75613
+[LPS-75624]: https://issues.liferay.com/browse/LPS-75624
 [LRDOCS-3023]: https://issues.liferay.com/browse/LRDOCS-3023